+++ Frames/SettingsEditors/BaseEditorFrame.lua
@@ -22,11 +22,36 @@ local Constants = _p.Constants;
 local FrameUtil = _p.FrameUtil;
 local OptionType = _p.Settings.OptionType;
 local PixelPerfect = _p.PixelPerfect;
+local ProfileManager = _p.ProfileManager;
 
 _p.BaseEditorFrame = {};
 local BaseEditorFrame = _p.BaseEditorFrame;
 
 local _constructors = {};
+
+local BaseEditorMixin = {};
+function BaseEditorMixin:SetOptionValue(...)
+    self.isChangingSettings = true;
+    self.option.Set(...);
+    self.isChangingSettings = false;
+end
+function BaseEditorMixin:IsChangingSettings()
+    return self.isChangingSettings;
+end
+function BaseEditorMixin:GetDefaultHeight()
+    return 30;
+end
+function BaseEditorMixin:SetDisabled(disabled)
+    if (self.option.IsActive == nil) then return false; end
+    if (self.disabled == disabled) then return false; end
+    self.disabled = disabled;
+    if (disabled) then
+        self.disabledBlocker:Show();
+    else
+        self.disabledBlocker:Hide();
+    end
+    return true;
+end
 
 function BaseEditorFrame.AddConstructor(type, func)
     if (_constructors[type] ~= nil) then error("Cannot register a type twice!"); end;
@@ -54,36 +79,12 @@ end
 function BaseEditorFrame.CreateRefreshSettingsFromProfile(handler)
     return function(self)
         self.isRefreshingFromProfile = true;
-        if (type(handler) == "table") then _p.Log(self); end;
         handler(self);
         self.isRefreshingFromProfile = false;
     end
 end
 
-local BaseEditorMixin = {};
-function BaseEditorMixin:SetOptionValue(...)
-    self.isChangingSettings = true;
-    self.option.Set(...);
-    self.isChangingSettings = false;
-end
-function BaseEditorMixin:IsChangingSettings()
-    return self.isChangingSettings;
-end
-function BaseEditorMixin:GetDefaultHeight()
-    return 30;
-end
-function BaseEditorMixin:SetDisabled(disabled)
-    if (self.option.IsActive == nil) then return false; end
-    if (self.disabled == disabled) then return false; end
-    self.disabled = disabled;
-    if (disabled) then
-        self.disabledBlocker:Show();
-    else
-        self.disabledBlocker:Hide();
-    end
-    return true;
-end
-
+BaseEditorFrame.Mixin = BaseEditorMixin;
 function BaseEditorFrame.Create(parent, option)
     if (_constructors[option.Type] == nil) then
         error("Couldn't find constructor for " .. option.Type);
@@ -93,6 +94,7 @@ end
 
 function BaseEditorFrame.CreateBaseFrame(parent, option)
     local frame = CreateFrame("Frame", nil, parent);
+    Mixin(frame, BaseEditorMixin);
     frame.option = option;
 
     if (frame.option.IsActive) then
@@ -105,10 +107,20 @@ function BaseEditorFrame.CreateBaseFrame(parent, option)
             FrameUtil.CreateTextTooltip(frame.disabledBlocker, option.Description, frame.disabledBlocker, "ANCHOR_LEFT", 0, 0, 1, 1, 1, 1);
         end
     end
-    
-
-    Mixin(frame, BaseEditorMixin);
     frame.isChangingSettings = false;
+
+    if (frame.option.ShouldRefreshOnProfileChanges) then
+        frame.OnProfilePropertyChanged = function()
+            frame:RefreshFromProfile();
+        end
+        ProfileManager.RegisterProfileChangedListener(function (newProfile, oldProfile)
+            if (oldProfile ~= nil) then
+                oldProfile:UnregisterAllPropertyChanged(frame.OnProfilePropertyChanged);
+            end
+            newProfile:RegisterAllPropertyChanged(frame.OnProfilePropertyChanged);
+        end);
+        ProfileManager:GetCurrent():RegisterAllPropertyChanged(frame.OnProfilePropertyChanged);
+    end
     return frame;
 end