+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # MacFrames Changelog
 
 ## v0.2.0 (2022/11/XX)
-* NEW: Choose from a variety of frame style presets or create your own.
+* NEW: Customize the frame colors freely or choose one of the provided presets (Default, Dark, Light).
 
 ## v0.1.7 (2022/11/06)
 * NEW: Sorting Raid Parties by Player Role